pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-ingestor"
version = "1.0.0"
description = "Phân hệ Tự động Thu thập và Xử lý Tin tức Tài chính - MCP Server Integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "click>=8.1",
    "streamlit>=1.37",
    "pandas>=2.2",
    "plotly>=5.23",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9",
    "google-genai>=1.0",
    "sentence-transformers>=3.0",
    "qdrant-client>=1.9",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "mcp>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29"]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
news-ingestor = "news_ingestor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q --tb=short"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["UP017"]

[tool.ruff.lint.per-file-ignores]
"dashboard.py" = ["E501"]
